# src/Legend2toolbox.web/src/api/generated/models/CleanUpCardRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CleanUpCardRequest = {
    cardId?: string | null;
    cdk?: string | null;
};