# src/Legend2toolbox.web/src/api/generated/models/CreateCardNumberRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateCardNumberRequest = {
    owner?: string | null;
    durationInDays?: number;
    faceValue?: number;
    amount?: number;
    notes?: string | null;
};