# src/Legend2toolbox.web/src/api/generated/models/ReissueCardRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ReissueCardRequest = {
    cardId?: string | null;
    cdk?: string | null;
};